requirements.txt
gradio
einops
Xformers
langchain
sentence-transformers
chromadb
unstructured
--extra-index-url https://download.pytorch.org/whl/cu118
torch
